requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

# ─── Núcleo científico ──────────────────────────────────────
numpy==1.26.4  # compatível com Label‑Studio (<2.0)
pandas>=2.2.3
scikit-learn==1.4.2
matplotlib==3.9.1
seaborn==0.13.2

# ─── NLP & embeddings ──────────────────────────────────────
nltk>=3.9.1
spacy==3.7.4
fasttext-wheel==0.9.2

# --- PyTorch GPU cu121 ---
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121

transformers==4.51.0
datasets==4.0.0

# ─── Scraping / coleta ────────────────────────────────────
praw==7.8.1
snscrape==0.7.0.20230622
tldextract>=5.1.3
python-dotenv==1.0.1
selenium==4.15.2
webdriver-manager==4.0.1

# ─── Rotulagem visual ─────────────────────────────────────
label-studio==1.20.0

# ─── Desenvolvimento / UX ─────────────────────────────────
jupyterlab==4.4.4
jupyterlab-git==0.40.0
rich==13.7.1
typer==0.9.0

# ─── Extras (opcionais) ───────────────────────────────────
wordcloud==1.9.4

# SpaCy modelo PT‑BR (download direto)
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0.tar.gz